# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Service",
        "description": "API by for longnguyen",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/information": {
            "get": {
                "tags": [
                    "Information"
                ],
                "summary": "Get List Information",
                "description": "Display a listing of the resource.",
                "operationId": "getInformation",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occurred while displaying a listing of the information."
                    }
                }
            }
        },
        "/api/v1/information/{id}": {
            "get": {
                "tags": [
                    "Information"
                ],
                "summary": "Get Information by ID",
                "description": "Display the specified resource.",
                "operationId": "getInformationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of information to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Information not found"
                    },
                    "500": {
                        "description": "An error occurred while displaying the information."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Information"
                ],
                "summary": "Update Information",
                "description": "Update Information",
                "operationId": "updateInformation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of information to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "information_name",
                                    "information_title",
                                    "information_phone",
                                    "information_email",
                                    "information_status",
                                    "information_opening_time",
                                    "information_closing_time",
                                    "information_mission",
                                    "information_description",
                                    "information_maps"
                                ],
                                "properties": {
                                    "information_name": {
                                        "type": "string",
                                        "example": "Barbershop"
                                    },
                                    "information_title": {
                                        "type": "string",
                                        "example": "Barbershop"
                                    },
                                    "information_phone": {
                                        "type": "string",
                                        "example": "01234567899"
                                    },
                                    "information_email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "information_images": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": ""
                                    },
                                    "information_status": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "information_opening_time": {
                                        "type": "string",
                                        "example": "08:00:00"
                                    },
                                    "information_closing_time": {
                                        "type": "string",
                                        "example": "17:00:00"
                                    },
                                    "information_mission": {
                                        "type": "string",
                                        "example": "Barbershop"
                                    },
                                    "information_description": {
                                        "type": "string",
                                        "example": "Barbershop"
                                    },
                                    "information_maps": {
                                        "type": "string",
                                        "example": "Barbershop"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Information not found"
                    },
                    "422": {
                        "description": "Invalid file."
                    },
                    "500": {
                        "description": "An error occurred while updating the information."
                    }
                }
            }
        },
...
    }
}